sys-apps/systemd: fix install logic that removes /sbin
Closes: https://bugs.gentoo.org/659042
Package-Manager: Portage-2.3.40_p15, Repoman-2.3.9_p247

+++ sys-apps/systemd/systemd-9999.ebuild
@@ -312,9 +312,12 @@ multilib_src_install_all() {
 
 	if ! use sysv-utils; then
 		rm "${ED%/}${rootprefix}"/sbin/{halt,init,poweroff,reboot,runlevel,shutdown,telinit} || die
-		rmdir "${ED%/}${rootprefix}"/sbin || die
 		rm "${ED%/}"/usr/share/man/man1/init.1 || die
 		rm "${ED%/}"/usr/share/man/man8/{halt,poweroff,reboot,runlevel,shutdown,telinit}.8 || die
+	fi
+
+	if ! use resolvconf && ! use sysv-utils; then
+		rmdir "${ED%/}${rootprefix}"/sbin || die
 	fi
 
 	# Preserve empty dirs in /etc & /var, bug #437008
